removed clutter files
