--- bam-ui/src/features/city/featuredCity.module.css
@@ -1,49 +1,0 @@
-.cities {
-    width: 100%;
-    max-width: 1024px;
-    display: flex;
-    justify-content: space-between;
-    gap: 10px;
-}
-.cities_item {
-    position: relative;
-    height :250px;
-    border-radius: 10px;
-    overflow: hidden;
-    padding-left: 5px;
-    cursor: pointer;
-    
-}
-.cities_img:hover {
-    transform: scale(1.1,1.1);
-    transition: 0.3s ease-in-out;
-}
-.cities_img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border-radius: 10px;
-    filter: brightness(80%) ;
-}
-.cities_title {
-    position: absolute;
-    bottom: 15px;
-    left:20px;
-    color: white;
-}
-.cities_title>h1{
-    margin-bottom: 5px;
-}
-.cities_title>h3{
-    margin-top: 5px;
-}
-.cities_secondary {
-    display: flex;
-    justify-content: space-evenly;
-    width: 100%;
-    gap: 20px;
-    max-width: 1024px;
-    position: relative;
-
-}
-
