--- bam-ui/src/features/property/propertyType.module.css
@@ -1,42 +1,0 @@
-.property_list {
-    display: flex;
-    max-width: 1024px;
-    justify-content: space-between;
-    gap: 10px;
-    border-radius: 5px;
-    overflow: hidden;
-    
-
-}
-.property_list_item{
-    cursor: pointer;
-    border-radius: 5px;
-    overflow: hidden;
-}
-
-
-.property_list_item_img {
-    width: 100%;
-    height: 160px;
-    object-fit: cover;
-    border-radius: 5px;
-}
-.property_list_item:hover {
-    transform: scale(1.1,1.1);
-    transition: 3ms ease-in-out;
-}
-.property_list_item_title {
-    font-size: 0.8rem;
-    padding-left: 2px;
-}
-.property_list_item_title>h1{
-    margin: 0px;
-    font-size: 22px;
-    
-}
-.property_list_item_title>h3{
-    margin-top: 0px;
-    font-size: 15px;
-    color: var(--gy);
-    
-}
