+++ bam-ui/src/forms/formlayout/LoginRegister.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import styles from "./layout.module.css";
+import styles from "./loginregister.module.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPersonThroughWindow } from "@fortawesome/free-solid-svg-icons";
 import UserLogin from "../login/UserLogin";
