+++ bam-ui/src/pages/propertyDetails/details.module.css
